# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.LessonCreate:
    properties:
      lesson_name:
        type: string
    type: object
  models.LessonResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      lesson_name:
        type: string
      topic_id:
        type: integer
      topic_name:
        type: string
      updated_at:
        type: string
    type: object
  models.LessonUpdate:
    properties:
      id:
        type: integer
      lesson_name:
        type: string
    type: object
...
info:
  contact: {}
  title: Admin-Monolight
paths:
...
  /v1/lesson:
    post:
      consumes:
      - application/json
      description: This Api for creating a new lesson
      parameters:
      - description: LessonCreate Model
        in: body
        name: LessonCreate
        required: true
        schema:
          $ref: '#/definitions/models.LessonCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LessonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Create Lesson
      tags:
      - lessons
    put:
      consumes:
      - application/json
      description: This Api for updating lesson
      parameters:
      - description: Update LessonUpdate Model
        in: body
        name: LessonUpdate
        required: true
        schema:
          $ref: '#/definitions/models.LessonUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LessonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Lesson
      tags:
      - lessons
  /v1/lesson/{id}:
    delete:
      consumes:
      - application/json
      description: This Api for deleting lesson
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete Lesson
      tags:
      - lessons
    get:
      consumes:
      - application/json
      description: This Api for get lesson
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LessonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Lesson
      tags:
      - lessons
  /v1/lessons:
    get:
      consumes:
      - application/json
      description: This Api for get all lessons
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LessonResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: ListLessons
      tags:
      - lessons
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"